requirements.txt
asgiref==3.7.2
colorama==0.4.6
distlib==0.3.7
django-deep-serializer==0.1.3
filelock==3.12.4
image==1.5.33
password-maker==0.0.3
pbr==5.11.1
Pillow==10.0.1
platformdirs==3.11.0
pypng==0.20220715.0
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
